Move merchants app enums to their own file
