+++ api/graphql/types/reservation/serializers/cancellation_serializers.py
@@ -1,4 +1,4 @@
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from graphene_django_extensions import NestingModelSerializer
 from rest_framework.exceptions import ValidationError
@@ -6,10 +6,13 @@ from rest_framework.exceptions import ValidationError
 from api.graphql.extensions import error_codes
 from common.date_utils import local_datetime
 from email_notification.helpers.reservation_email_notification_sender import ReservationEmailNotificationSender
-from merchants.models import OrderStatus, PaymentOrder
+from merchants.enums import OrderStatus
 from reservations.choices import ReservationStateChoice
 from reservations.models import Reservation
 from reservations.tasks import refund_paid_reservation_task
+
+if TYPE_CHECKING:
+    from merchants.models import PaymentOrder
 
 __all__ = [
     "ReservationCancellationSerializer",
