+++ api/graphql/types/reservation/serializers/refund_serializers.py
@@ -5,7 +5,8 @@ from django.utils.timezone import get_default_timezone
 
 from api.graphql.extensions.serializers import OldPrimaryKeySerializer
 from api.graphql.extensions.validation_errors import ValidationErrorCodes, ValidationErrorWithCode
-from merchants.models import OrderStatus, PaymentOrder
+from merchants.enums import OrderStatus
+from merchants.models import PaymentOrder
 from reservations.choices import ReservationStateChoice
 from reservations.models import Reservation
 from reservations.tasks import refund_paid_reservation_task
