+++ reservations/querysets/reservation.py
@@ -8,7 +8,7 @@ from django.db import models
 from helsinki_gdpr.models import SerializableMixin
 
 from common.date_utils import local_datetime
-from merchants.models import OrderStatus
+from merchants.enums import OrderStatus
 from reservation_units.models import ReservationUnit
 from reservations.choices import ReservationStateChoice
