+++ tests/factories/reservation.py
@@ -5,7 +5,7 @@ from typing import Any
 from factory import fuzzy
 
 from common.date_utils import local_start_of_day, next_hour
-from merchants.models import OrderStatus, PaymentType
+from merchants.enums import OrderStatus, PaymentType
 from reservation_units.enums import PricingType
 from reservation_units.models import ReservationUnit
 from reservations.choices import ReservationStateChoice, ReservationTypeChoice
