+++ tests/test_external_services/test_verkkokauppa/test_webhooks/test_order_payment_webhooks.py
@@ -4,7 +4,7 @@ import pytest
 from django.urls import reverse
 
 from email_notification.helpers.reservation_email_notification_sender import ReservationEmailNotificationSender
-from merchants.models import OrderStatus
+from merchants.enums import OrderStatus
 from merchants.verkkokauppa.payment.exceptions import GetPaymentError
 from merchants.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from reservations.choices import ReservationStateChoice
