+++ tests/test_graphql_api/test_reservation/test_delete.py
@@ -2,7 +2,7 @@ import uuid
 
 import pytest
 
-from merchants.models import OrderStatus
+from merchants.enums import OrderStatus
 from merchants.verkkokauppa.order.exceptions import CancelOrderError
 from merchants.verkkokauppa.verkkokauppa_api_client import VerkkokauppaAPIClient
 from reservations.choices import ReservationStateChoice
