+++ tests/test_helauth/test_gdpr_api.py
@@ -7,7 +7,7 @@ from dateutil.relativedelta import relativedelta
 from django.urls import reverse
 from django.utils import timezone
 
-from merchants.models import OrderStatus
+from merchants.enums import OrderStatus
 from reservations.choices import ReservationStateChoice
 from tests.factories import ApplicationFactory, PaymentOrderFactory, ReservationFactory, UserFactory
 from tests.test_helauth.helpers import get_gdpr_auth_header, patch_oidc_config
